+++ src/navigation/index.js
@@ -5,7 +5,7 @@
  */
 import { registerBlockType } from '@wordpress/blocks'
 import { useBlockProps, InnerBlocks, InspectorControls } from '@wordpress/block-editor'
-import { PanelBody, SelectControl, ToggleControl } from '@wordpress/components'
+import { PanelBody, SelectControl } from '@wordpress/components'
 
 /**
  * Retrieves the translation of text.
@@ -30,6 +30,13 @@ const customIcon = createElement(
 		}
 	)
 )
+
+/**
+ * Función auxiliar: random para Ids
+ */
+function getRandomArbitrary (min, max) {
+	return Math.floor(Math.random() * (max - min) + min)
+}
 
 /**
  * Lets webpack process CSS, SASS or SCSS files referenced in JavaScript files.
@@ -93,7 +100,7 @@ registerBlockType('ekiline-block-collection/ekiline-navbar', {
 		anchor: true,
 		color: {
 			gradients: true // Enables the gradients UI control.
-		}
+		},
 	},
 
 	/**
@@ -112,18 +119,21 @@ registerBlockType('ekiline-block-collection/ekiline-navbar', {
 			type: 'string',
 			default: ' navbar-expand-lg' // lg,md,sm.
 		},
-		alignToggler: {
-			type: 'boolean',
-			default: false // markup before-after.
-		},
 		alignItems: {
 			type: 'string',
 			default: '', // justify-content-md-center
-		},
-		navMenu: {
-			type: 'string',
-			default: '' //id de navegacion *select.
 		}
+	},
+	/**
+	 * Se ocupara contexto para pasar valores.
+	 * @link https://developer.wordpress.org/block-editor/reference-guides/block-api/block-context/
+	 */
+	providesContext: {
+		'ekiline-navbar/anchor': 'anchor',
+		'ekiline-navbar/navPosition': 'navPosition',
+		'ekiline-navbar/navStyle': 'navStyle',
+		'ekiline-navbar/navShow': 'navShow',
+		'ekiline-navbar/alignItems': 'alignItems'
 	},
 	/**
 	 * @see ./edit.js
@@ -132,6 +142,7 @@ registerBlockType('ekiline-block-collection/ekiline-navbar', {
 	edit: (props) => {
 
 		const { attributes, setAttributes } = props
+
 		// personalizar clase
 		const blockProps = useBlockProps({
 			className: 'editor-navbar'
@@ -142,22 +153,21 @@ registerBlockType('ekiline-block-collection/ekiline-navbar', {
 			setAttributes({ anchor: 'customNav' + getRandomArbitrary(10, 150) })
 		}
 
-		// Add new classname validating navPosition attribute
-		if (attributes.navPosition) {
-			blockProps.className += attributes.navPosition
-		}
-		if (attributes.navShow) {
-			blockProps.className += attributes.navShow
-		}
-		if (attributes.alignItems) {
-			blockProps.className += attributes.alignItems
-		}
-
+		// Restringir los bloques.
+		const PARENT_ALLOWED_BLOCKS = [
+			'core/site-logo',
+			'core/site-title',
+			'core/paragraph',
+			'core/buttons',
+			'ekiline-block-collection/ekiline-navbar-menu-wrapper',
+			'ekiline-block-collection/ekiline-navbar-toggler'
+		]
 		// Cargar un preset.
 		const CHILD_TEMPLATE = [
-			['core/site-logo'],
-			['core/paragraph', { placeholder: __('Add a block', 'ekiline-block-collection') }],
-			['core/navigation']
+			['core/site-logo',{"width":60,"className":"navbar-brand mb-0"}],
+			['core/site-title',{"className":"navbar-brand mb-0","style":{"typography":{"fontSize":"16px"}}}],
+			['ekiline-block-collection/ekiline-navbar-toggler'],
+			['ekiline-block-collection/ekiline-navbar-menu-wrapper']
 		]
 
 		// Función separada para determinar el texto de ayuda
@@ -222,28 +232,6 @@ registerBlockType('ekiline-block-collection/ekiline-navbar', {
 						help={getHelpText(attributes.navShow)}
 					/>
 
-					<ToggleControl
-						label={__('Nav button align', 'ekiline-block-collection')}
-						checked={attributes.alignToggler}
-						onChange={(alignToggler) =>
-								setAttributes({ alignToggler })}
-						help={
-							(!attributes.alignToggler)
-							? __('Default right', 'ekiline-block-collection')
-							:__('Align left', 'ekiline-block-collection')
-						}
-					/>
-
-					<SelectControl
-						label={__('Select Menu', 'ekiline-block-collection')}
-						value={attributes.navMenu}
-						options={[
-							{ label: __('Default', 'ekiline-block-collection'), value: '' },
-						]}
-						onChange={(navMenu) =>
-							setAttributes({ navMenu })}
-					/>
-
 					<SelectControl
 						label={__('Align nav items', 'ekiline-block-collection')}
 						value={attributes.alignItems}
@@ -260,8 +248,9 @@ registerBlockType('ekiline-block-collection/ekiline-navbar', {
 
 			{/* El bloque */}
 			<InnerBlocks
-				orientation='horizontal'
+				allowedBlocks={PARENT_ALLOWED_BLOCKS}
 				template={CHILD_TEMPLATE}
+				orientation='horizontal'
 			/>
 		</div>
 		)
@@ -276,7 +265,6 @@ registerBlockType('ekiline-block-collection/ekiline-navbar', {
 		const blockProps = useBlockProps.save({
 			className: 'navbar',
 		})
-
 		// Add new classname validating navPosition attribute
 		if (attributes.navPosition) {
 			blockProps.className += attributes.navPosition
@@ -288,71 +276,10 @@ registerBlockType('ekiline-block-collection/ekiline-navbar', {
 			blockProps.className += attributes.alignItems
 		}
 
-		// Componente Boton.
-		function NavToggler () {
-			// Toggler classnames
-			let togglerClassnames = (attributes.navPosition === ' fixed-bottom' || attributes.navPosition === ' sticky-bottom') ? 'navbar-toggler order-5' : 'navbar-toggler';
-			// Toggle alinear botón
-			togglerClassnames += (!attributes.alignToggler) ? ' ms-auto' : '';
-
-			return (
-				<button
-					className={togglerClassnames}
-					type='button'
-					data-bs-toggle={attributes.navStyle}
-					data-bs-target={'#' + attributes.anchor + 'Child'}
-					aria-controls={attributes.anchor + 'Child'}
-					aria-expanded={(attributes.navStyle !== 'offcanvas')?'false':null}
-					aria-label={__('Toggle navigation', 'ekiline-block-collection')}>
-						<span class='dashicons dashicons-menu'></span>
-				</button>
-			)
-		}
-
-		// atributos para collapse
-		const collapseProps = {
-			id: attributes.anchor + 'Child',
-			className: (attributes.navStyle === 'offcanvas') ? 'offcanvas offcanvas-end' : attributes.navStyle + ' navbar-collapse'
-		}
-		if (attributes.alignItems) {
-			collapseProps['className'] += attributes.alignItems
-		}
-
-		// Componente Boton cerrar (offcanvas)
-        const BtnClose = () => {
-            if (attributes.navStyle !== 'offcanvas') return null
-            const addClassnamesBtn = (data) => {
-                switch (data) {
-                    case ' navbar-expand-lg':
-                        return ' d-lg-none'
-                    case ' navbar-expand-sm':
-                        return ' d-sm-none'
-                    case ' navbar-expand':
-                        return ' d-none'
-                    default:
-                        return ''
-                }
-            }
-            return (
-                <button
-                    type='button'
-                    className={'btn-close' + addClassnamesBtn(attributes.navShow)}
-                    data-bs-dismiss='offcanvas'
-                    aria-label='Close'
-                ></button>
-            )
-        }
-
 		return (
 			<div {...blockProps}>
 				<div className='container-fluid'>
-					<NavToggler />
-					<div>{'contenido a modificar'}</div>
-					<div {...collapseProps}>
-						<BtnClose />
-						<InnerBlocks.Content />
-					</div>
-
+					<InnerBlocks.Content/>
 				</div>
 			</div>
 		)
@@ -360,8 +287,231 @@ registerBlockType('ekiline-block-collection/ekiline-navbar', {
 })
 
 /**
- * Función auxiliar: random para Ids
+ * Bloque interno navbar
+ * - nav.navbar + navbar-expand-lg + bg-body-tertiary [fixed-top/fixed-bottom/sticky-top/sticky-bottom]
+ * - - div + container-fluid + [container-md/sm/]
+ * - - - a.navbar-brand [href]
+ * - - - button.navbar-toggler [+ data]
+ * - - - - span .navbar-toggler-icon
+ * - - div + collapse navbar-collapse [+ id] variables off canvas [offcanvas offcanvas-end +  otros valores]
+ * - - - - [bloque navegacion UL] intervenir clases css + [navbar-nav-scroll + style="--bs-scroll-height: 100px;"]
+ * - - - - [Permitir otros bloques] intervenir clases css // Funciona para offcanvas.
  */
-function getRandomArbitrary (min, max) {
-	return Math.floor(Math.random() * (max - min) + min)
-}
+registerBlockType('ekiline-block-collection/ekiline-navbar-menu-wrapper', {
+	title: __('Navbar menu container', 'ekiline-block-collection'),
+	parent: ['ekiline-block-collection/ekiline-navbar'],
+	icon: customIcon,
+	description: __('Add items to your navigation bar', 'ekiline-block-collection'),
+	category: 'design',
+	// Se ocupa contexto para pasar valores desde el padre, en este caso el ID.
+	usesContext: [
+		'ekiline-navbar/anchor',
+		'ekiline-navbar/navStyle',
+		'ekiline-navbar/alignItems',
+		'ekiline-navbar/navShow',
+		'ekiline-navbar/navPosition'
+	],
+	supports: {
+		anchor: true,
+		html: false, // no permitir HTML
+		reusable: false,
+	},
+	attributes: {
+		parentAnchor: {
+		type: 'string',
+		default: '' // remove dataset [data-bs-parent].
+		},
+		parentNavStyle: {
+			type: 'string',
+			default: ' collapse' // offcanvas,nav-scroller
+		},
+		parentAlignItems: {
+			type: 'string',
+			default: '', // justify-content-md-center
+		},
+		parentNavShow: {
+			type: 'string',
+			default: '' // lg,md,sm.
+		},
+		parentNavPosition: {
+			type: 'string',
+			default: '' // fixed-top, fixed-bottom, sticky-top, sticky-bottom.
+		}
+	},
+	/**
+	 * @see ./edit.js
+	 */
+	// edit: Edit,
+	edit: (props) => {
+
+		const { attributes, setAttributes } = props
+
+		// Cargar un preset.
+		const CHILD_TEMPLATE = [
+			['core/navigation',{"overlayMenu":"never","className":"navbar-nav","style":{"typography":{"fontSize":"16px"}}}]
+		]
+
+		// personalizar clase
+		const blockProps = useBlockProps({
+			className: 'editor-collapse editor-navbar-collapse',
+		})
+
+		// Precargar nombre ID en hijos y valores heredados de contexto.
+		if (!attributes.parentAnchor) {
+			setAttributes({ parentAnchor: props.context['ekiline-navbar/anchor'] })
+		}
+
+		// Precargar nombre ID (anchor).
+		if (!attributes.anchor) {
+			setAttributes({ anchor: props.context['ekiline-navbar/anchor'] + 'Child' })
+		}
+
+		// Actualizar estado parentNavStyle.
+		setAttributes({ parentNavStyle: props.context['ekiline-navbar/navStyle'] })
+		// Actualizar estado parentAlignItems.
+		setAttributes({ parentAlignItems: props.context['ekiline-navbar/alignItems'] })
+		// Actualizar estado parentNavShow.
+		setAttributes({ parentNavShow: props.context['ekiline-navbar/navShow'] })
+		// Actualizar estado parentNavPosition.
+		setAttributes({ parentNavPosition: props.context['ekiline-navbar/navPosition'] })
+
+		return (
+			<InnerBlocks
+				orientation='horizontal'
+				template={CHILD_TEMPLATE}
+			/>
+		)
+	},
+
+	/**
+	 * @see ./save.js
+	 */
+	// save,
+	save: ({ attributes }) => {
+
+		// Modificar className de bloque.
+		function addClassnames(data) {
+			if (data === 'offcanvas') {
+				return 'offcanvas offcanvas-end'
+			}
+			return data + ' navbar-collapse';
+		}
+
+		const blockProps = useBlockProps.save({
+			className: addClassnames(attributes.parentNavStyle),
+			'data-bs-parent': (attributes.parentAlignItems && attributes.parentAnchor) ? '#' + attributes.parentAnchor : null
+		})
+		// En caso de fixed-bottom o sticky-bottom.
+		if (attributes.parentNavPosition === ' fixed-bottom' || attributes.parentNavPosition === ' sticky-bottom') {
+			blockProps.className += ' order-first'
+		}
+
+		// Modificar className de boton.
+		function addClassnamesBtn(data) {
+			switch (data) {
+				case ' navbar-expand-lg':
+					return ' d-lg-none'
+				case ' navbar-expand-sm':
+					return ' d-sm-none'
+				case ' navbar-expand':
+					return ' d-none'
+				default:
+					return ''
+			}
+		}
+
+		return (
+			<div {...blockProps}>
+				{attributes.parentNavStyle === 'offcanvas' && (
+					<button type="button" 
+						class={'btn-close' + addClassnamesBtn(attributes.parentNavShow) } 
+						data-bs-dismiss="offcanvas" 
+						aria-label="Close"></button>
+				)}
+				<InnerBlocks.Content />
+			</div>
+		)
+	}
+
+})
+
+/**
+ * Bloque interno toggler, componente.
+ */
+registerBlockType('ekiline-block-collection/ekiline-navbar-toggler', {
+	title: __('Navbar toggler', 'ekiline-block-collection'),
+	parent: ['ekiline-block-collection/ekiline-navbar'],
+	icon: customIcon,
+	description: __('Add a button to toggle the navigation bar', 'ekiline-block-collection'),
+	category: 'design',
+	// Se ocupa contexto para pasar valores desde el padre, en este caso el ID.
+	usesContext: [
+		'ekiline-navbar/anchor',
+		'ekiline-navbar/navStyle',
+		'ekiline-navbar/navPosition',
+	],
+	supports: {
+		anchor: true,
+		html: false,
+		color: {
+			text: false
+		},
+	},
+	attributes: {
+		parentAnchor: {
+		type: 'string',
+		default: '' // remove dataset [data-bs-parent].
+		},
+		parentNavStyle: {
+			type: 'string',
+			default: ' collapse' // offcanvas,nav-scroller
+		}
+	},
+	/**
+	 * @see ./edit.js
+	 */
+	// edit: Edit,
+	edit: (props) => {
+
+		const { attributes, setAttributes } = props
+
+		// personalizar clase
+		const blockProps = useBlockProps({
+			className: 'editor-navbar-toggler'
+		})
+		// Precargar nombre ID en hijos y valores heredados de contexto.
+		if (!attributes.parentAnchor) {
+			setAttributes({ parentAnchor: props.context['ekiline-navbar/anchor'] })
+		}
+		// Precargar nombre ID (anchor).
+		if (!attributes.anchor) {
+			setAttributes({ anchor: props.context['ekiline-navbar/anchor'] + 'Toggler' })
+		}
+		// Actualizar estado parentNavStyle.
+		setAttributes({ parentNavStyle: props.context['ekiline-navbar/navStyle'] })
+
+		return (
+			<button {...blockProps}>
+				<span class='dashicons dashicons-menu'></span>
+			</button>
+		)
+	},
+	save: ({ attributes }) => {
+
+		const blockProps = useBlockProps.save({
+			className: 'navbar-toggler',
+			'data-bs-toggle': attributes.parentNavStyle,
+			'data-bs-target': '#' + attributes.parentAnchor + 'Child',
+			'aria-controls': attributes.parentAnchor + 'Child',
+			'aria-expanded': (attributes.parentNavStyle !== 'offcanvas')?'false':null,
+			'aria-label': __('Toggle navigation', 'ekiline-block-collection')
+		})
+
+		return (
+			<button {...blockProps}>
+				<span class="navbar-toggler-icon"></span>
+			</button>
+		)
+	}
+});
+
